Refactoring | cleaning up Util, moving methods to its specific classes
