+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/model/elastic/converters/DynamicFieldsConverter.java
@@ -1,13 +1,17 @@
 package org.broadinstitute.dsm.model.elastic.converters;
 
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 
+import org.apache.commons.lang3.StringUtils;
 import org.broadinstitute.dsm.model.elastic.Util;
 import org.broadinstitute.dsm.model.elastic.export.parse.BaseParser;
 import org.broadinstitute.dsm.model.elastic.export.parse.DynamicFieldsParser;
 import org.broadinstitute.dsm.model.elastic.export.parse.ValueParser;
 import org.broadinstitute.dsm.statics.ESObjectConstants;
+import org.broadinstitute.dsm.util.proxy.jackson.ObjectMapperSingleton;
 
 public class DynamicFieldsConverter extends BaseConverter {
 
@@ -20,9 +24,14 @@ public class DynamicFieldsConverter extends BaseConverter {
     }
 
     @Override
+    public void setParser(BaseParser parser) {
+        this.parser = parser;
+    }
+
+    @Override
     public Map<String, Object> convert() {
         Map<String, Object> finalResult;
-        Map<String, Object> objectMap = Util.dynamicFieldsSpecialCase(fieldValue);
+        Map<String, Object> objectMap = dynamicFieldsSpecialCase(fieldValue);
         Map<String, Object> transformedMap = new HashMap<>();
         for (Map.Entry<String, Object> object : objectMap.entrySet()) {
             String field = object.getKey();
@@ -37,8 +46,38 @@ public class DynamicFieldsConverter extends BaseConverter {
         return finalResult;
     }
 
-    @Override
-    public void setParser(BaseParser parser) {
-        this.parser = parser;
+    private Map<String, Object> dynamicFieldsSpecialCase(Object fieldValue) {
+        Map<String, Object> dynamicMap = new HashMap<>();
+        if (isJsonInString(fieldValue)) {
+            String strValue = (String) fieldValue;
+            try {
+                dynamicMap = ObjectMapperSingleton.instance().readValue(strValue, Map.class);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        }
+        return dynamicMap;
+    }
+
+    private boolean isJsonInString(Object fieldValue) {
+        return fieldValue instanceof String && StringUtils.isNotBlank((String) fieldValue) && isJson((String) fieldValue);
+    }
+
+    private boolean isJson(String str) {
+        return getFirstChar(str) == '{' && getLastChar(str) == '}';
+    }
+
+    private char getLastChar(String strValue) {
+        if (Objects.isNull(strValue) || strValue.length() == 0) {
+            throw new IllegalArgumentException();
+        }
+        return strValue.charAt(strValue.length() - 1);
+    }
+
+    private char getFirstChar(String strValue) {
+        if (Objects.isNull(strValue) || strValue.length() == 0) {
+            throw new IllegalArgumentException();
+        }
+        return strValue.charAt(0);
     }
 }
