+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/model/elastic/export/process/SingleProcessor.java
@@ -1,9 +1,12 @@
 package org.broadinstitute.dsm.model.elastic.export.process;
 
+import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 
-import org.broadinstitute.dsm.model.elastic.Util;
+import com.fasterxml.jackson.core.type.TypeReference;
+import org.broadinstitute.dsm.util.proxy.jackson.ObjectMapperSingleton;
 
 public class SingleProcessor extends BaseProcessor {
 
@@ -17,7 +20,9 @@ public class SingleProcessor extends BaseProcessor {
 
     @Override
     protected Map<String, Object> convertObjectToCollection(Object object) {
-        return Util.convertObjectToMap(object);
+        return Objects.isNull(object)
+                ? new HashMap<>()
+                : ObjectMapperSingleton.instance().convertValue(object, new TypeReference<Map<String, Object>>() {});
     }
 
     @Override
