+++ pepper-apis/dsm-core/src/test/java/org/broadinstitute/dsm/model/elastic/search/SourceMapDeserializerTest.java
@@ -1,16 +1,47 @@
 package org.broadinstitute.dsm.model.elastic.search;
 
+import static org.junit.Assert.assertEquals;
+
 import java.io.IOException;
+import java.lang.reflect.Field;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
+import org.broadinstitute.dsm.model.FollowUp;
 import org.broadinstitute.dsm.statics.ESObjectConstants;
 import org.broadinstitute.dsm.util.proxy.jackson.ObjectMapperSingleton;
 import org.junit.Assert;
 import org.junit.Test;
 
 public class SourceMapDeserializerTest {
+
+    @Test
+    public void getParameterizedType() throws NoSuchFieldException {
+        class MockClass {
+            List<Object> listField;
+            FollowUp[] followUps;
+            Object obj;
+        }
+
+        Field listField = MockClass.class.getDeclaredField("listField");
+        Field followUps = MockClass.class.getDeclaredField("followUps");
+        Field obj = MockClass.class.getDeclaredField("obj");
+
+        SourceMapDeserializer sourceMapDeserializer = new SourceMapDeserializer();
+
+        Class<?> clazz = null;
+        try {
+            clazz = sourceMapDeserializer.getParameterizedType(listField.getGenericType());
+            assertEquals(Object.class, clazz);
+            clazz = sourceMapDeserializer.getParameterizedType(followUps.getGenericType());
+            assertEquals(FollowUp.class, clazz);
+            clazz = sourceMapDeserializer.getParameterizedType(obj.getGenericType());
+            assertEquals(Object.class, clazz);
+        } catch (ClassNotFoundException e) {
+            Assert.fail();
+        }
+    }
 
     @Test
     public void convertFollowUpsJsonToList() {
